backend/avg_speed_service/requirements.txt
fastapi
uvicorn
redis
kafka-python
python-dotenv
